+++ generate-models.ts
@@ -1,0 +1,1 @@
+se risolve il problema si